Move netcdf file to utils
